Refactorings and changes
YouTube has a Loop button on the right click menu of the video.
Unfortunately, while that sets the loop attribute on the video element,
it does not "observe" it for changes.  So just setting the loop like
that would not change the menu.  That was an annoyance so instead of
setting loop directly, we'll instead fake click on the menu.  May break
in the future but whatever.