Port to Gio
Instead of using pydbus, just use Gio directly.  Frankly the only reason
that was done was to allow Properties to be async, rather than the
currect hack of "query and pray".  This does allows to drop a dependency
though, and that's nice.

TBH, this does a bit more than it should, they other thing is that it
splits the MediaPlayer2 class up.  The idea is that each provider will
have its own subclass that will discribe the "passive" capabilities of
the site, those than won't change between videos.  Things are still a
bit hardcoded for YouTube perhaps, but if/when I add more sites I'll
refactor that.