Do Position the proper way
Though it's proper it's so so ugly... I'm starting to think that the
previous hack may even have been better...  In any case, no
functionallity should have changed.

Also, handle all video IDs that contain -.  These are valid on YouTube
but not as DBus object paths.  So we do some weird escaping.